refactor: Chat tile and Conversation list improvements
- Renamed UserTileWidget to ChatTileWidget for clarity
- Updated ChatTileWidget to use the new user identifier param
- Refactored ChatTIleWidget  Widget structure and updated padding values
- Moved curentUser logic from ChatTileWidget to ConversationListPage so the curent user is ignored from chat list
