--- lib/view/widgets/user_tile.dart
@@ -1,67 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:pandasenp_flutter/controllers/auth.dart';
-import 'package:pandasenp_flutter/model/user.dart';
-import 'package:pandasenp_flutter/view/pages/chat.dart';
-
-class UserTileWidget extends StatelessWidget {
-  UserTileWidget({required this.user, super.key});
-  final User user;
-  final AuthController auth = AuthController();
-  final anonyme = 'Anonyme';
-
-  @override
-  Widget build(BuildContext context) {
-    return FutureBuilder<User>(
-      future: auth.currentUser,
-      builder: (context, snapshot) {
-        if (snapshot.connectionState == ConnectionState.none ||
-            !snapshot.hasData) {
-          return const Center(
-            child: CircularProgressIndicator(),
-          );
-        } else if (snapshot.hasError) {
-          return Center(
-            child: Text(
-              snapshot.error.toString(),
-              style: const TextStyle(
-                fontSize: 20,
-              ),
-            ),
-          );
-        }
-        User currentUser = snapshot.data!;
-        return GestureDetector(
-          onTap: () {
-            if (currentUser.id == user.id) return;
-            Navigator.push(
-              context,
-              MaterialPageRoute(
-                builder: (context) => ChatPage(
-                  recipient: user,
-                ),
-              ),
-            );
-          },
-          child: Padding(
-            padding: const EdgeInsets.only(top: 8.0),
-            child: Card(
-              color: Colors.black.withOpacity(0.7),
-              margin: const EdgeInsets.only(
-                  top: 12.0, bottom: 6.0, left: 20.0, right: 20.0),
-              child: ListTile(
-                title: Text(
-                  user.firstName ?? user.email ?? anonyme,
-                  style: const TextStyle(color: Colors.white),
-                ),
-                subtitle: const Text(
-                  "dernière envoi de la conversation",
-                  style: TextStyle(color: Colors.grey),
-                ),
-              ),
-            ),
-          ),
-        );
-      },
-    );
-  }
-}
